+++ NEWS.md
@@ -6,6 +6,7 @@ Find out the main highlights of each update.
 
 Fixes:
   * Prevent crash if all detections for a given stray tag were removed during detection quality checks.
+  * Prevent migration crash if all movement events are rendered invalid before section movements are created.
 
 Enhancements:
   * the shapefiles of the study areas can now be water or land polygons. Use the new argument "time" in loadShape() to switch from land shapefiles (the default) to water shapefiles.